1. Add PollParentDto in AddPollDto, removed toChatSpacePoll() and toStreamPoll() method. 2. Refactored and updated AddPollDto to make working parent polls easier. 3. Refactored AddReviewDto and added ReviewParentDto and removed toStreamReview(). 4. Changed Review likeCount type from Long to Integer. 5. Replaced likeCount in ChatSpaceResponse with LikeCountInfo. 6. Updated ChatSpaceResponse and add LikeCountInfo. 7. Updated usages of FleenStream.getTotalAttendees() by replacing variable assignment type to Integer. 8. Updated review table schema in db.changelog-create-schema.sql by replacing member_id with author_id. 9. Moved method from ReviewServiceImpl to FleenStream for checking user eligibility to add review if the stream has not yet started. 10. Updated JoinStreamResponse by replacing totalAttending Long type with Integer type. 11. Updated messages_en_US.properties for info classes like LikeInfo. 12. Removed addChatSpacePoll() and addStreamPoll() from PollController. 13. Updated PollServiceImpl.addPoll() to unify and making easy to add polls with or without parent. 14. Updated PollMapperImpl to set organizer id to allow for updating the poll eligibility for update by the requesting user. 15. Added isUpdatable in PollResponse and implemented Updatable interface. 16. Added new static method in PollResponseEntriesHolder. 17. Updated PollSearchController by setting authentication principal user variable on PollSearchController.findPoll(). 18. Updated PollSearchService and PollSearchServiceImpl. 19. Removed methods in PollService and PollServiceImpl. 20. Updated ProcessAttendeeRequestToJoinStreamResponse, RequestToJoinStreamResponse and RetrieveStreamResponse to replace Long type with Integer type in totalAttending member variable. 21. Renamed "member" and "memberId" variable in Review to author and authorId. 22. Added Open API/Swagger documentation on ReviewController. 23. Updated ReviewOtherDetailsHolder to add getter method for getting parentTitle. 24. Added exceptions thrown in method signature of ReviewService.addReview() and ReviewService.updateReview(). 25. Updated ReviewServiceImpl.updateReview() and .addReview() due to change in AddReviewDto. 26. Added LikeCount and LikeCountInfo in StreamResponse. 27. Added ReviewCount and ReviewCountInfo. 28. Created AddPollParentDetailHolder for holding parent details before creating poll.

+++ src/main/java/com/fleencorp/feen/model/domain/chat/ChatSpace.java
@@ -87,7 +87,7 @@ public class ChatSpace extends FleenFeenEntity {
   private Set<Link> links = new HashSet<>();
 
   @Column(name = "like_count", nullable = false)
-  private Long likeCount = 0L;
+  private Integer likeCount = 0;
 
   public Member getOrganizer() {
     return member;
