+++ src/main/java/com/fleencorp/feen/model/response/stream/attendance/RequestToJoinStreamResponse.java
@@ -37,7 +37,7 @@ public class RequestToJoinStreamResponse extends LocalizedResponse {
   private StreamTypeInfo streamTypeInfo;
 
   @JsonProperty("total_attending")
-  private Long totalAttending;
+  private Integer totalAttending;
 
   @JsonIgnore
   protected StreamType getStreamType() {
@@ -49,7 +49,7 @@ public class RequestToJoinStreamResponse extends LocalizedResponse {
     return StreamType.isEvent(getStreamType()) ? "request.to.join.event" : "request.to.join.live.broadcast";
   }
 
-  public static RequestToJoinStreamResponse of(final Long streamId, final AttendanceInfo attendanceInfo, final StreamTypeInfo streamTypeInfo, final Long totalAttending) {
+  public static RequestToJoinStreamResponse of(final Long streamId, final AttendanceInfo attendanceInfo, final StreamTypeInfo streamTypeInfo, final Integer totalAttending) {
     return new RequestToJoinStreamResponse(streamId, attendanceInfo, streamTypeInfo, totalAttending);
   }
 }
